Add chisel-util/chiselc
This package installs a wrapper for scalac.  According to
https://github.com/ducky64/chiselc

  Chisel compiler wrapper scripts

  THIS IS ALPHA-STAGE AND EXPERIMENTAL SOFTWARE. Use at your own risk;
  interfaces may change.

This script is going to be necessary to install any Chisel packages.

"chisel-util" is a new category, this patch adds it to "categories".
The general idea here is that "chisel-util" is meant to keep utilities
related to Chisel that are not themselves Chisel libraries.